.ipynb_checkpoints/requirements-checkpoint.txt
torch
torchvision
streamlit
pillow
numpy
pandas
matplotlib
seaborn
scikit-learn
